refactor: split holdings responsibilities
- move ETF catalog, source loading, and output writing into focused modules

- leave holdings.py as the public API and CLI orchestrator

- update tests and agent notes for the new layout
