+++ src/index_etfs/holdings.py
@@ -1,303 +1,19 @@
-"""📊 Extract ticker symbols from various index ETFs.
+"""📊 Extract ticker symbols from index ETFs."""
 
-This module downloads ETF/index holdings from public providers (SSGA, iShares, Nasdaq)
-and extracts just the ticker symbols (e.g., AAPL, MSFT, GOOGL).
-"""
-
-import io
-import json
-import math
-import urllib.request
-from dataclasses import dataclass
-from datetime import UTC, datetime
 from pathlib import Path
-from typing import Literal
 
 import polars as pl
 
-
-Provider = Literal["ssga", "ishares", "nasdaq"]
-
-
-@dataclass(frozen=True)
-class ETFConfig:
-    url: str
-    provider: Provider
-    watchlist_name: str
-    expected_count: int
-
-
-# 🎯 ETF Configuration
-ETF_CONFIGS = {
-    "spy": ETFConfig(
-        url="https://www.ssga.com/us/en/individual/library-content/products/fund-data/etfs/us/holdings-daily-us-en-spy.xlsx",
-        provider="ssga",
-        watchlist_name="sp500",
-        expected_count=503,
-    ),
-    "mdy": ETFConfig(
-        url="https://www.ssga.com/us/en/individual/library-content/products/fund-data/etfs/us/holdings-daily-us-en-mdy.xlsx",
-        provider="ssga",
-        watchlist_name="sp400",
-        expected_count=400,
-    ),
-    "spsm": ETFConfig(
-        url="https://www.ssga.com/us/en/individual/library-content/products/fund-data/etfs/us/holdings-daily-us-en-spsm.xlsx",
-        provider="ssga",
-        watchlist_name="sp600",
-        expected_count=600,
-    ),
-    "qqq": ETFConfig(
-        url="https://api.nasdaq.com/api/quote/list-type/nasdaq100",
-        provider="nasdaq",
-        watchlist_name="nasdaq100",
-        expected_count=100,
-    ),
-    "iwm": ETFConfig(
-        url="https://www.blackrock.com/varnish-api/blk-one01-product-data/product-data/api/v1/get-fund-document?appType=PRODUCT_PAGE&appSubType=ISHARES&targetSite=us-ishares&locale=en_US&portfolioId=239710&userType=individual&component=holdings",
-        provider="ishares",
-        watchlist_name="russell2000",
-        expected_count=2000,
-    ),
-}
-
-FIREFOX_HEADERS = {
-    "User-Agent": "Mozilla/5.0 (X11; Linux x86_64; rv:140.0) Gecko/20100101 Firefox/140.0",
-    "Accept": "*/*",
-    "Accept-Language": "en-US,en;q=0.5",
-}
-
-TRADINGVIEW_SCAN_URL = "https://scanner.tradingview.com/america/scan"
-MIN_EXPECTED_RATIO = 0.9
-
-
-def _validate_count(symbol: str, count: int) -> None:
-    """Fail closed on empty or suspiciously small source data."""
-    if count == 0:
-        raise ValueError(f"{symbol.upper()} returned zero rows")
-
-    expected = ETF_CONFIGS[symbol].expected_count
-    minimum = math.ceil(expected * MIN_EXPECTED_RATIO)
-    if count < minimum:
-        raise ValueError(
-            f"{symbol.upper()} returned {count} rows; expected at least {minimum}"
-        )
-
-
-def _write_metadata(results: dict[str, int], output_dir: Path | None = None) -> None:
-    """Write latest generated counts and source URLs."""
-    if output_dir is None:
-        output_dir = Path.cwd()
-
-    metadata = {
-        "generated_at": datetime.now(UTC)
-        .replace(microsecond=0)
-        .isoformat()
-        .replace("+00:00", "Z"),
-        "watchlists": {
-            config.watchlist_name: {
-                "count": count,
-                "expected_count": config.expected_count,
-                "minimum_count": math.ceil(config.expected_count * MIN_EXPECTED_RATIO),
-                "source": symbol,
-                "source_url": config.url,
-            }
-            for symbol, count in results.items()
-            for config in [ETF_CONFIGS[symbol]]
-        },
-    }
-
-    metadata_dir = output_dir / "metadata"
-    metadata_dir.mkdir(parents=True, exist_ok=True)
-    (metadata_dir / "latest.json").write_text(json.dumps(metadata, indent=2) + "\n")
-
-
-def _read_url(url: str) -> io.BytesIO:
-    """Read a URL with browser-ish headers."""
-    request = urllib.request.Request(url, headers=FIREFOX_HEADERS)
-    with urllib.request.urlopen(request, timeout=30) as response:
-        return io.BytesIO(response.read())
-
-
-def _tradingview_symbols(tickers: list[str], chunk_size: int = 200) -> dict[str, tuple[str, str]]:
-    """Map tickers to TradingView symbols and sector-ish group names."""
-    symbols = {}
-    headers = FIREFOX_HEADERS | {
-        "Content-Type": "application/json",
-        "Origin": "https://www.tradingview.com",
-        "Referer": "https://www.tradingview.com/",
-    }
-
-    for start in range(0, len(tickers), chunk_size):
-        chunk = tickers[start : start + chunk_size]
-        payload = {
-            "filter": [{"left": "name", "operation": "in_range", "right": chunk}],
-            "columns": ["name", "sector", "industry"],
-            "markets": ["america"],
-        }
-        request = urllib.request.Request(
-            TRADINGVIEW_SCAN_URL,
-            data=json.dumps(payload).encode(),
-            headers=headers,
-        )
-        with urllib.request.urlopen(request, timeout=30) as response:
-            rows = json.load(response).get("data", [])
-
-        for row in rows:
-            data = row.get("d", [])
-            ticker = data[0] if data else None
-            symbol = row.get("s", "")
-            group = next((value for value in data[1:3] if value), "Other")
-            if ticker in chunk and ":" in symbol:
-                symbols.setdefault(ticker, (symbol, group))
-
-    return symbols
-
-
-def _watchlist_lines(tickers: list[str], symbols: dict[str, tuple[str, str]]) -> list[str]:
-    """Build TradingView watchlist lines grouped by ### headers."""
-    groups: dict[str, list[str]] = {}
-    for ticker in tickers:
-        symbol, group = symbols.get(ticker, (ticker, "Other"))
-        groups.setdefault(group, []).append(symbol)
-
-    return [line for group in sorted(groups) for line in (f"###{group}", *groups[group])]
-
-
-def _load_ssga_excel(url: str) -> pl.DataFrame:
-    """🏢 Load SSGA Excel format holdings data.
-
-    Args:
-        url: URL to the SSGA Excel file
-
-    Returns:
-        Polars DataFrame with holdings data
-    """
-    df_dict = pl.read_excel(_read_url(url), sheet_id=0, read_options={"header_row": 4})
-
-    # 🎲 Handle both dict and DataFrame returns
-    if isinstance(df_dict, dict):
-        df = list(df_dict.values())[0]
-    else:
-        df = df_dict
-
-    return df
-
-
-def _load_ishares_csv(url: str) -> pl.DataFrame:
-    """🏢 Load iShares CSV format holdings data.
-
-    Args:
-        url: URL to the iShares CSV file
-
-    Returns:
-        Polars DataFrame with holdings data
-    """
-    df = pl.read_csv(
-        _read_url(url),
-        skip_rows=9,
-        columns=["Ticker", "Name", "Weight (%)", "Market Currency"],
-        truncate_ragged_lines=True,
-    )
-    df = df.rename({"Weight (%)": "Weight"})
-    return df
-
-
-def _load_nasdaq_json(url: str) -> pl.DataFrame:
-    """🏢 Load Nasdaq-100 symbols from Nasdaq's public JSON endpoint."""
-    rows = json.load(_read_url(url)).get("data", {}).get("data", {}).get("rows")
-    if not rows:
-        raise ValueError("Nasdaq response did not contain any Nasdaq-100 rows")
-
-    tickers = [row["symbol"].strip().replace("/", ".") for row in rows if row.get("symbol")]
-    if not tickers:
-        raise ValueError("Nasdaq response did not contain any symbols")
-
-    return pl.DataFrame({"Ticker": tickers})
-
-
-def _filter_and_clean(df: pl.DataFrame, provider: str) -> pl.DataFrame:
-    """🧹 Filter and clean holdings data based on provider format.
-
-    Args:
-        df: Input DataFrame
-        provider: ETF provider name (ssga, ishares, nasdaq)
-
-    Returns:
-        Cleaned DataFrame with just Ticker column (uppercase symbols only)
-    """
-    # 🏃 Handle empty DataFrame early - just select Ticker column
-    if len(df) == 0:
-        return df.select("Ticker") if "Ticker" in df.columns else pl.DataFrame({"Ticker": []})
-
-    if provider in ("ssga", "ishares"):
-        currency_col = "Local Currency" if provider == "ssga" else "Market Currency"
-        df = df.filter((pl.col(currency_col) == "USD") & (pl.col("Ticker") != "-"))
-
-    # 🎯 Filter out null/empty tickers, special placeholders, get uppercase symbols only, sorted alphabetically
-    df = df.filter(
-        pl.col("Ticker").is_not_null()
-        & (pl.col("Ticker") != "")
-        & (pl.col("Ticker") != "-")
-        & ~pl.col("Ticker").str.contains("_", literal=True)  # Exclude CASH_USD and similar placeholders
-        & ~pl.col("Ticker").str.contains(" ", literal=True)  # Exclude warrants like "BBBY WS"
-    ).select("Ticker").sort("Ticker")
-    return df
-
-
-def _save_holdings(df: pl.DataFrame, symbol: str, output_dir: Path | None = None) -> None:
-    """💾 Save ticker symbols to a text file.
-
-    Args:
-        df: DataFrame containing ticker symbols
-        symbol: ETF symbol (used for filename)
-        output_dir: Optional output directory (defaults to current directory)
-    """
-    if output_dir is None:
-        output_dir = Path.cwd()
-
-    output_dir.mkdir(parents=True, exist_ok=True)
-
-    # 📝 Save as simple newline-separated list of tickers
-    tickers = df["Ticker"].to_list()
-    tickers_dir = output_dir / "tickers"
-    tickers_dir.mkdir(exist_ok=True)
-    (tickers_dir / f"{symbol}.txt").write_text("\n".join(tickers) + "\n")
-
-    config = ETF_CONFIGS.get(symbol)
-    if config is None:
-        return
-
-    watchlist_dir = output_dir / "watchlists"
-    watchlist_dir.mkdir(exist_ok=True)
-    symbols = _tradingview_symbols(tickers)
-    missing = [ticker for ticker in tickers if ticker not in symbols]
-    if missing:
-        print(f"⚠️  No TradingView metadata for {', '.join(missing)}")
-    (watchlist_dir / f"{config.watchlist_name}.txt").write_text(
-        "\n".join(_watchlist_lines(tickers, symbols)) + "\n"
-    )
+from .catalog import ETF_CONFIGS, validate_count
+from .outputs import save_holdings, write_metadata
+from .sources import filter_and_clean, load_holdings
 
 
 def get_etf_holdings(
     symbol: str,
     output_dir: Path | None = None,
 ) -> pl.DataFrame:
-    """📈 Get ticker symbols for a specific ETF.
-
-    This is the main function for downloading and extracting ticker symbols.
-    It handles different provider formats automatically.
-
-    Args:
-        symbol: ETF symbol (spy, mdy, spsm, qqq, iwm)
-        output_dir: Optional directory to save output file
-
-    Returns:
-        DataFrame with just the Ticker column (uppercase symbols, sorted alphabetically)
-
-    Raises:
-        ValueError: If the ETF symbol is not recognized
-    """
+    """Download, clean, save, and return holdings for one configured ETF."""
     symbol_lower = symbol.lower()
 
     if symbol_lower not in ETF_CONFIGS:
@@ -307,88 +23,51 @@ def get_etf_holdings(
         )
 
     config = ETF_CONFIGS[symbol_lower]
-    loaders = {
-        "ssga": _load_ssga_excel,
-        "ishares": _load_ishares_csv,
-        "nasdaq": _load_nasdaq_json,
-    }
-    try:
-        loader = loaders[config.provider]
-    except KeyError as exc:
-        raise ValueError(f"Unknown provider: {config.provider}") from exc
-
-    # 🔄 Load data based on provider
-    df = loader(config.url)
-
-    # 🧹 Clean and filter
-    df = _filter_and_clean(df, config.provider)
-    _validate_count(symbol_lower, len(df))
-
-    # 💾 Save outputs
-    _save_holdings(df, symbol_lower, output_dir)
-
+    df = filter_and_clean(load_holdings(config), config.provider)
+    validate_count(symbol_lower, len(df))
+    save_holdings(df, symbol_lower, output_dir)
     return df
 
 
 def get_spy_holdings() -> pl.DataFrame:
-    """📊 Get ticker symbols from the SPY ETF.
-
-    Returns:
-        DataFrame with SPY ticker symbols
-    """
+    """📊 Get ticker symbols from the SPY ETF."""
     return get_etf_holdings("spy")
 
 
 def get_mdy_holdings() -> pl.DataFrame:
-    """📊 Get ticker symbols from the MDY ETF.
-
-    Returns:
-        DataFrame with MDY ticker symbols
-    """
+    """📊 Get ticker symbols from the MDY ETF."""
     return get_etf_holdings("mdy")
 
 
 def get_spsm_holdings() -> pl.DataFrame:
-    """📊 Get ticker symbols from the SPSM ETF.
-
-    Returns:
-        DataFrame with SPSM ticker symbols
-    """
+    """📊 Get ticker symbols from the SPSM ETF."""
     return get_etf_holdings("spsm")
 
 
 def get_qqq_holdings() -> pl.DataFrame:
-    """📊 Get ticker symbols from the Nasdaq-100.
-
-    Returns:
-        DataFrame with QQQ ticker symbols
-    """
+    """📊 Get ticker symbols from the Nasdaq-100."""
     return get_etf_holdings("qqq")
 
 
 def get_iwm_holdings() -> pl.DataFrame:
-    """📊 Get ticker symbols from the IWM ETF.
-
-    Returns:
-        DataFrame with IWM ticker symbols
-    """
+    """📊 Get ticker symbols from the IWM ETF."""
     return get_etf_holdings("iwm")
 
 
 def main() -> None:
-    """🚀 Main function to download all ETF holdings."""
+    """🚀 Download all ETF holdings."""
     print("📥 Downloading ETF holdings...")
     output_dir = Path.cwd()
     results = {}
 
-    for symbol in ETF_CONFIGS.keys():
+    for symbol in ETF_CONFIGS:
         print(f"  ⏳ Processing {symbol.upper()}...")
         df = get_etf_holdings(symbol, output_dir)
         count = len(df)
         results[symbol] = count
         print(f"  ✅ {symbol.upper()} complete! ({count} rows)")
 
-    _write_metadata(results, output_dir)
+    write_metadata(results, output_dir)
     print("🎉 All holdings downloaded successfully!")
